Extract compression types from bogo::main
